Create task module to save BC pipeline utils
