+++ dea_burn_cube/algo.py
@@ -1,8 +1,6 @@
 import ctypes
-import datetime
 import logging
 import multiprocessing as mp
-import time
 from contextlib import closing
 
 import numpy as np
@@ -14,113 +12,10 @@ from shapely import geometry
 from shapely.geometry import Point
 from shapely.ops import unary_union
 
+import dea_burn_cube.task as task
+
 logger = logging.getLogger(__name__)
 logging.basicConfig(level=logging.INFO, format="%(asctime)s %(message)s")
-
-
-def log_execution_time(func):
-    def wrapper(*args, **kwargs):
-        start_time = time.time()
-        result = func(*args, **kwargs)
-        end_time = time.time()
-        duration = end_time - start_time
-        logging.info(f"{func.__name__} took {duration:.2f} seconds to execute.")
-        return result
-
-    return wrapper
-
-
-def task_to_ranges(task_id, task_table):
-
-    _ = s3fs.S3FileSystem(anon=True)
-
-    periods_columns = [
-        "Period Start",
-        "Period End",
-        "Mapping Period Start",
-        "Mapping Period End",
-    ]
-
-    task_map = pd.read_csv(
-        f"s3://dea-public-data-dev/projects/burn_cube/configs/{task_table}",
-        parse_dates=periods_columns,
-        dayfirst=True,
-    )
-
-    result_dict = {}
-
-    if task_id not in list(task_map["Processing Name"]):
-        return result_dict
-
-    task_info = task_map[task_map["Processing Name"] == task_id].iloc[0]
-
-    for periods_column in periods_columns:
-        result_dict[periods_column] = task_info[periods_column].strftime("%Y-%m-%d")
-
-    return result_dict
-
-
-def dynamic_task_to_ranges(dtime):
-    import calendar
-
-    period_last_day = calendar.monthrange(dtime.year - 1, dtime.month)[1]
-
-    result_dict = {}
-
-    # TODO: change me if we plan to move to 3/2 years processing period
-    result_dict["Period Start"] = datetime.datetime(
-        dtime.year - 5, dtime.month + 1, 1
-    ).strftime("%Y-%m-%d")
-    result_dict["Period End"] = datetime.datetime(
-        dtime.year - 1, dtime.month, period_last_day
-    ).strftime("%Y-%m-%d")
-
-    mapping_period_last_day = calendar.monthrange(dtime.year, dtime.month)[1]
-    result_dict["Mapping Period Start"] = datetime.datetime(
-        dtime.year - 1, dtime.month + 1, 1
-    ).strftime("%Y-%m-%d")
-    result_dict["Mapping Period End"] = datetime.datetime(
-        dtime.year, dtime.month, mapping_period_last_day
-    ).strftime("%Y-%m-%d")
-
-    return result_dict
-
-
-def generate_task(task_id, task_table):
-    updated_task_id = f"{task_id.split('-')[0]}-20{task_id.split('-')[1]}"
-    dtime = datetime.datetime.strptime(updated_task_id, "%b-%Y")
-
-    if dtime.year > 2023:
-        result_dict = dynamic_task_to_ranges(dtime)
-    else:
-        result_dict = task_to_ranges(task_id, task_table)
-    return result_dict
-
-
-def stretch_rgb(data):
-    """
-    stretch RGB to 0-1
-    """
-    tmp = data
-    a = (
-        (tmp - np.nanpercentile(tmp, 1))
-        / (np.nanpercentile(tmp, 99.5) - np.nanpercentile(tmp, 0.5))
-        * 255
-    )
-    a[a > 255] = 255
-    a[a < 0] = 0
-    return a
-
-
-def rgb_plot(r, g, b):
-    from PIL import Image
-
-    rgb_array = np.zeros((r.shape[0], r.shape[1], 3), "uint8")
-    rgb_array[..., 0] = stretch_rgb(r)
-    rgb_array[..., 1] = stretch_rgb(g)
-    rgb_array[..., 2] = stretch_rgb(b)
-    img = Image.fromarray(np.flipud(rgb_array))
-    return img.resize((500, 500))
 
 
 def cos_distance(ref, obs):
@@ -142,8 +37,8 @@ def cos_distance(ref, obs):
     cosdist = np.transpose(
         1
         - np.nansum(ref * obs, axis=0)
-        / np.sqrt(np.sum(ref**2))
-        / np.sqrt(np.nansum(obs**2, axis=0))
+        / np.sqrt(np.sum(ref ** 2))
+        / np.sqrt(np.nansum(obs ** 2, axis=0))
     )
     return cosdist
 
@@ -166,7 +61,7 @@ def nbr_eucdistance(ref, obs):
     nbr_dist.fill(np.nan)
     index = np.where(~np.isnan(obs))[0]
     euc_dist = obs[index] - ref
-    euc_norm = np.sqrt(euc_dist**2)
+    euc_norm = np.sqrt(euc_dist ** 2)
     nbr_dist[index] = euc_norm
     direction[index[euc_dist < -0.05]] = 1
 
@@ -524,7 +419,7 @@ def dist_severity(params):
         )
 
 
-@log_execution_time
+@task.log_execution_time
 def distances(ard, geomed):
     """
     Calculates the cosine distance between observation and reference.
@@ -672,7 +567,7 @@ def distances(ard, geomed):
     return ds
 
 
-@log_execution_time
+@task.log_execution_time
 def outliers(dataset, distances):
     """
     Calculate the outliers for distances for change detection
@@ -889,7 +784,7 @@ def create_attributes(dataset, product_name, version, method, res=30):
     return dataset
 
 
-@log_execution_time
+@task.log_execution_time
 def hotspot_polygon(period, extent, buffersize, hotspotfile):
     """Create polygons for the hotspot with a buffer
     year: given year for hotspots data
@@ -964,7 +859,7 @@ def hotspot_polygon(period, extent, buffersize, hotspotfile):
     return polygons
 
 
-@log_execution_time
+@task.log_execution_time
 def severitymapping(
     dists, outlrs, period, hotspotfile, method="NBR", growing=True, hotspots_period=None
 ):
