+++ dea_burn_cube/task.py
@@ -1,0 +1,88 @@
+import datetime
+import logging
+import time
+
+import pandas as pd
+import s3fs
+
+logger = logging.getLogger(__name__)
+logging.basicConfig(level=logging.INFO, format="%(asctime)s %(message)s")
+
+
+def log_execution_time(func):
+    def wrapper(*args, **kwargs):
+        start_time = time.time()
+        result = func(*args, **kwargs)
+        end_time = time.time()
+        duration = end_time - start_time
+        logging.info(f"{func.__name__} took {duration:.2f} seconds to execute.")
+        return result
+
+    return wrapper
+
+
+def task_to_ranges(task_id, task_table):
+
+    _ = s3fs.S3FileSystem(anon=True)
+
+    periods_columns = [
+        "Period Start",
+        "Period End",
+        "Mapping Period Start",
+        "Mapping Period End",
+    ]
+
+    task_map = pd.read_csv(
+        f"s3://dea-public-data-dev/projects/burn_cube/configs/{task_table}",
+        parse_dates=periods_columns,
+        dayfirst=True,
+    )
+
+    result_dict = {}
+
+    if task_id not in list(task_map["Processing Name"]):
+        return result_dict
+
+    task_info = task_map[task_map["Processing Name"] == task_id].iloc[0]
+
+    for periods_column in periods_columns:
+        result_dict[periods_column] = task_info[periods_column].strftime("%Y-%m-%d")
+
+    return result_dict
+
+
+def dynamic_task_to_ranges(dtime):
+    import calendar
+
+    period_last_day = calendar.monthrange(dtime.year - 1, dtime.month)[1]
+
+    result_dict = {}
+
+    # TODO: change me if we plan to move to 3/2 years processing period
+    result_dict["Period Start"] = datetime.datetime(
+        dtime.year - 5, dtime.month + 1, 1
+    ).strftime("%Y-%m-%d")
+    result_dict["Period End"] = datetime.datetime(
+        dtime.year - 1, dtime.month, period_last_day
+    ).strftime("%Y-%m-%d")
+
+    mapping_period_last_day = calendar.monthrange(dtime.year, dtime.month)[1]
+    result_dict["Mapping Period Start"] = datetime.datetime(
+        dtime.year - 1, dtime.month + 1, 1
+    ).strftime("%Y-%m-%d")
+    result_dict["Mapping Period End"] = datetime.datetime(
+        dtime.year, dtime.month, mapping_period_last_day
+    ).strftime("%Y-%m-%d")
+
+    return result_dict
+
+
+def generate_task(task_id, task_table):
+    updated_task_id = f"{task_id.split('-')[0]}-20{task_id.split('-')[1]}"
+    dtime = datetime.datetime.strptime(updated_task_id, "%b-%Y")
+
+    if dtime.year > 2023:
+        result_dict = dynamic_task_to_ranges(dtime)
+    else:
+        result_dict = task_to_ranges(task_id, task_table)
+    return result_dict
